+++ features/notifications/ticket-strategy/src/main/java/org/opennms/netmgt/notifd/TicketNotificationStrategy.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2012-2014 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2014 The OpenNMS Group, Inc.
+ * Copyright (C) 2012-2022 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -151,7 +151,7 @@ public class TicketNotificationStrategy implements NotificationStrategy {
         }
         
         // We know the event is an alarm, pull the alarm and current ticket details from the database
-        AlarmState alarmState = getAlarmStateFromEvent(Integer.parseInt(eventID));
+        AlarmState alarmState = getAlarmStateFromEvent(Long.parseLong(eventID));
         if( alarmState.getAlarmID() == 0 ) {
 		LOG.error("There is no alarm-id associated with the event-id='{}'. Will not create ticket.", eventID);
         	return 1;
@@ -172,7 +172,7 @@ public class TicketNotificationStrategy implements NotificationStrategy {
      *
      * @return 0 if alarmid is null
      */
-	protected AlarmState getAlarmStateFromEvent(int eventID) {
+	protected AlarmState getAlarmStateFromEvent(long eventID) {
 		AlarmStateRowCallbackHandler callbackHandler = new AlarmStateRowCallbackHandler();
 
         JdbcTemplate template = new JdbcTemplate(DataSourceFactory.getInstance());
