+++ features/opennms-es-rest/src/main/java/org/opennms/plugins/elasticsearch/rest/EventToIndex.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2002-2014 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2014 The OpenNMS Group, Inc.
+ * Copyright (C) 2018-2022 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -222,8 +222,8 @@ public class EventToIndex implements AutoCloseable {
 	private Index createEventIndexFromEvent(final Event event) {
 		final JSONObject body = new JSONObject();
 
-		final Integer id = event.getDbid();
-		body.put("id", Integer.toString(id));
+		final Long id = event.getDbid();
+		body.put("id", Long.toString(id));
 		body.put("eventuei", event.getUei());
 
 		final Calendar cal=Calendar.getInstance();
@@ -307,7 +307,7 @@ public class EventToIndex implements AutoCloseable {
 		// document to the event's database ID. Otherwise, allow ES to
 		// generate a unique ID value.
 		if (id != null && id > 0) {
-			builder = builder.id(Integer.toString(id));
+			builder = builder.id(Long.toString(id));
 		}
 
 		Index index = builder.build();
