+++ opennms-ackd/src/test/java/org/opennms/netmgt/ackd/AckdIT.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2009-2020 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2020 The OpenNMS Group, Inc.
+ * Copyright (C) 2009-2022 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -385,7 +385,7 @@ public class AckdIT implements InitializingBean {
     }
 
     class VerificationObject {
-        int m_eventID;
+        long m_eventID;
         int m_alarmId;
         int m_nodeId;
         int m_notifId;
