+++ opennms-services/src/main/java/org/opennms/netmgt/poller/QueryManager.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2004-2014 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2014 The OpenNMS Group, Inc.
+ * Copyright (C) 2004-2022 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2022 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -61,7 +61,7 @@ public interface QueryManager {
     /**
      * Set or updates the lost event id on the specified outage.
      */
-    void updateOpenOutageWithEventId(int outageId, int lostEventId);
+    void updateOpenOutageWithEventId(int outageId, long lostEventId);
 
     /**
      * Marks the outage for the given service as resolved
@@ -75,7 +75,7 @@ public interface QueryManager {
     /**
      * Set or updates the regained event id on the specified outage.
      */
-    void updateResolvedOutageWithEventId(int outageId, int regainedEventId);
+    void updateResolvedOutageWithEventId(int outageId, long regainedEventId);
 
     /**
      * @param nodeId
@@ -85,11 +85,11 @@ public interface QueryManager {
 
     void closeOutagesForUnmanagedServices();
 
-    void closeOutagesForNode(Date closeDate, int eventId, int nodeId);
+    void closeOutagesForNode(Date closeDate, long eventId, int nodeId);
 
-    void closeOutagesForInterface(Date closeDate, int eventId, int nodeId, String ipAddr);
+    void closeOutagesForInterface(Date closeDate, long eventId, int nodeId, String ipAddr);
 
-    void closeOutagesForService(Date closeDate, int eventId, int nodeId, String ipAddr, String serviceName);
+    void closeOutagesForService(Date closeDate, long eventId, int nodeId, String ipAddr, String serviceName);
 
     void updateServiceStatus(int nodeId, String ipAddr, String serviceName, String status);
